basic slider completed, needs styling
