+++ src/routes/home/home.component.jsx
@@ -6,8 +6,8 @@ import image3 from "../../assets/DSC_7018.jpg";
 import image4 from "../../assets/DSC_7022.jpg";
 import image5 from "../../assets/DSC_7032.jpg";
 
-import { BsChevronCompactLeft, BsChevronCompactRight } from "react-icons/bs";
-import { RxDotFilled } from "react-icons/rx";
+
+import Carousel from "../../components/carousel/Carousel";
 
 // https://www.youtube.com/watch?v=tXlZCW26bto
 
@@ -30,57 +30,10 @@ export default function Home() {
     },
   ];
 
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const prevSlide = () => {
-    const isFirstSlide = currentIndex === 0;
-    const newIndex = isFirstSlide ? slides.length - 1 : currentIndex - 1;
-    setCurrentIndex(newIndex);
-  };
-
-  const nextSlide = () => {
-    const isLastSlide = currentIndex === slides.length - 1;
-    const newIndex = isLastSlide ? 0 : currentIndex + 1;
-    setCurrentIndex(newIndex);
-  };
-
-  const goToSlide = (slideIndex) => {
-    setCurrentIndex(slideIndex);
-  };
-
   return (
-    <div>
-    <div className="max-w-screen h-[780px] w-screen m-auto relative group">
-      <div
-        style={{ backgroundImage: `url(${slides[currentIndex].url})` }}
-        className="w-full h-full bg-center bg-cover duration-500"
-      ></div>
-      {/* left arrow */}
-      <div className="hidden group-hover:block text-white absolute top-[50%] translate-x-0 translate-y-[-50%] left-5 text-2xl rounded-full p-2 bg-black/20  cursor-pointer">
-        <BsChevronCompactLeft onClick={prevSlide} size={30} />
-      </div>
-
-      {/* right arrow */}
-      <div className="hidden group-hover:block text-white absolute top-[50%] translate-x-0 translate-y-[-50%] right-5 text-2xl rounded-full p-2 bg-black/20  cursor-pointer">
-        <BsChevronCompactRight onClick={nextSlide} size={30} />
-      </div>
-
-      {/* dots */}
-      <div className="absolute bottom-0 left-[50%] transform translate-x-1/2">
-      <div className="flex top-4 justify-center py-2 text-white">
-        {slides.map((slide, slideIndex) => (
-          <div
-            key={slideIndex}
-            onClick={() => goToSlide(slideIndex)}
-            className="text-2xl cursor-pointer"
-          >
-            <RxDotFilled />
-          </div>
-        ))}
-      </div>
-      </div>
-    </div>
-    <a
+    <>
+      <Carousel autoSlide={false} slides={slides}></Carousel>
+      <a
         className="App-link"
         href="https://reactjs.org"
         target="_blank"
@@ -88,6 +41,6 @@ export default function Home() {
       >
         Learn React
       </a>
-    </div>
+    </>
   );
 }
